CDRIVER-2192 refactor cursor option handling
mongoc_collection_aggregate, in particular, was prone to merging
options into the command body twice. And when talking with mongos, it
was very complex to handle $query, $readPreference, and $clusterTime
properly. The logic that merges options into a command to produce a
command cursor is now in one place, and the functions that do so for
mongos and mongod are now separated.